Add end city chest loot
It's worth noting that with this commit, the repository includes YAML
configs for all vanilla chest loot tables. You can use j2cli and
loot_table.j2 to generate JSON files that are identical to the
vanilla loot tables except for an extra newline at the end of the
file (or just run `make lootoverhaul.zip`).