Overhaul armorer chest loot
In vanilla, toolsmith and weaponsmith chests have loot tables that
roll 3-8 times, and the weaponsmith table includes all four types
of wearable armor. Meanwhile, armorer chests roll 1-5 times and the
only armor entry is for a helmet.

Redo the armorer chest table to be based on the weaponsmith table.
The original armorer table contains emerald and no obsidian, so
use emerald instead of obsidian on the new table. Also, replace
the sword and axe entries (both weight 5) with a shield entry
(weight 10).