Adjust Sleep Overhaul pack
Advancements can't detect when a player hits an end crystal. So
instead of spawning phantoms whenever a crystal is damaged, spawn
Mother Phantoms when the dragon is summoned, and regular phantoms
when a Mother Phantom takes damage. Mother Phantoms have a chance
to drop a head and/or an elytra.