Redo Otherside loot
Make Otherside obtainable from creepers that are killed by skeletons.

Since nonspecific #music_discs loot is already capable of generating
Otherside, redo its entries in stronghold corridors and dungeons.

This caused be to rearrange a bunch of entries in those tables. The
overall effect is that, in stronghold corridors, bricks and ender
eyes are slightly more common and music discs are rarer. In dungeons,
petrified slabs, mob heads, and enchanted shields are rarer, and
enchanted iron swords, axes, and shovels now generate.