Override loot using datapack ordering instead of scoreboards
This requires more effort from the user since they have to put
the packs in the correct order, but it's less buggy and easier to
maintain.

In the future I plan on documenting the ordering requirements better
and numbering the zip files so that the priority is obvious.

In addition, fix an issue where the ancient city loot table wasn't
being included in Loot Overhaul.