Rename escape_end.zip to escape_the_end.zip
If this pack and escape_nether.zip are installed at the same time,
I *think* this will make it so that you play minecraft in reverse,
i.e. you have to beat the end and then the nether in order to get
to the overworld.

I need to test though.