Add broken Frienderchest implementation
For some reason, the exit function only works sometimes. When it
doesn't, it teleports the player to the overworld without changing
their coords.

I think a way to address this might be to use two advancements.
The first would trigger when the player enters an end gateway
and teleport them to the overworld. The second would trigger when
the player changes dimensions from the Frienderchest to the
overworld, and would teleport the player to their saved exit
location.