Frienderchest: Work around /tp bug
For some reason, teleporting from the frienderchest to the overworld
occasionally doesn't change the player's coordinates correctly,
so they end up teleporting to 1, 0, 1 in the overworld. I believe
this is a bug that's beyond a datapack's ability to fix, so as a
workaround, create an air hole at that location so that the player
doesn't suffocate.